+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "client",
+  "name": "path-explorer",
   "version": "0.1.0",
   "private": true,
   "scripts": {
