+++ client/src/styles/globals.css
@@ -1,8 +1,11 @@
 @import "tailwindcss";
 
 :root {
-  --background: #ffffff;
-  --foreground: #171717;
+  /* --background: #ffffff;
+  --foreground: #171717; */
+  --background: linear-gradient(180deg, #7a28b6 0.29%, #390d58 100%);
+  --foreground: #ededed;
+  background-color: var(--background);
 }
 
 @theme inline {
@@ -14,8 +17,9 @@
 
 @media (prefers-color-scheme: dark) {
   :root {
-    --background: #0a0a0a;
-    --foreground: #ededed;
+    /* --background: #0a0a0a; */
+    --foreground: #ededed; 
+    --background: linear-gradient(180deg, #7a28b6 0.29%, #390d58 100%);
   }
 }
